Store AliJaddi/requirements.txt
streamlit>=1.28.0
streamlit-option-menu>=0.3.2
pandas>=1.5.0
bcrypt>=4.0.1
pyotp>=2.8.0
cryptography>=41.0.0
PyJWT>=2.8.0
flask>=2.2.0
flask-cors>=4.0.0
requests>=2.31.0
python-dotenv>=1.0.0
plotly>=5.14.0
Pillow>=10.0.0
pytest>=7.4.0
